+++ NEWS.md
@@ -29,7 +29,8 @@
 
 - In `CreateWebMap` function, remove coordinates and zoom level information from top of map.
 
-- Add `AddRefreshButton` and `AddClusterButton` functions, used to add miscellaneous web map buttons.
+- Add `AddRefreshButton`, `AddClusterButton`, and `AddSearchButton` functions,
+  used to add miscellaneous web map buttons.
 
 - In `FindOptimalSubset` function, allow integer chromosomes to be specified for the `suggestions` argument.